Adds further condition in DmsfQuery#dmsf_nodes
It will make sure that a pending link will be
deleted from the query and hence not displayed
to the user.

A pending link has no project assigned. For those
links the case statement in DmsfQuery#dmsf_nodes
will now return true what causes the link to be
deleted from the query result.