Supports option -d to set index database via CLI
For running xapian indexer it is necessary to configure it to use the same database for indexing as defined in the user interface on plugin settings.

In order to make the configuration of xapian indexer some kind of persistence the command line options in `redmine_dmsf/extra/xapian_indexer.rb` could be extended by providing a further option for the index database path:

```shell
ruby redmine_dmsf/extra/xapian_indexer.rb -d <path/to/my/custom/index/database>
```

With this additional option it is possible to move the index database outside of redmine root to keep it even after running updates. Furthermore, custom configuration in `redmine_dmsf/extra/xapian_indexer.rb` won't be overridden when updating this plugin.