Adds further text to reset button of webdav digest
When a user has 2FA enabled the WebDAV digest won't be created on sign
in. The reason is in RedmineDmsf::Hooks::Controllers::
AccountControllerHooks#controller_account_success_authentication_after
where the digest will be created only if the controller parameter
':password' is present. This works for a user authentication with login
name and password only. A user with 2fa enabled runs differently through
the authentication process and crosses the hook not before the 2fa
token was checked. Hence, there won't be a password parameter anymore.

Instead of manipulating controller params to provide the password
only the button text for reseting the digest will be changed if the
user has 2FA but no digest yet. This would make it more explicit that
the token is not expected to exist and can be created if missing.